# openapi.yaml
openapi: 3.0.3
info:
  title: caceasy API
  version: 0.1.0
  description: API for coupon scanning and wallet management
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminApiKey:
      type: apiKey
      in: header
      name: x-admin-key
  schemas:
...
    Coupon:
      type: object
      properties:
        id:
          type: string
        token:
          type: string
        batch_id:
          type: string
          format: uuid
        points:
          type: integer
        status:
          type: string
          enum: [issued, redeemed]
        created_at:
          type: string
          format: date-time
        redeemed_at:
          type: string
          format: date-time
          nullable: true
      required:
        - token
        - batch_id
        - points
        - status
...
    Wallet:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        points:
          type: integer
        rupeeEquivalent:
          type: integer
      required:
        - user_id
        - points
        - rupeeEquivalent
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
paths:
...
  /coupons/generate:
    post:
      summary: Generate coupons for a batch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                batch_id:
                  type: string
                  format: uuid
                quantity:
                  type: integer
                  minimum: 1
                  maximum: 10000
                points:
                  type: integer
                  minimum: 1
                prefix:
                  type: string
              required:
                - batch_id
                - quantity
                - points
      responses:
        '201':
          description: Coupons generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  batch_id:
                    type: string
                  count:
                    type: integer
                  coupons:
                    type: array
                    items:
                      type: object
                      properties:
                        token:
                          type: string
                        batch_id:
                          type: string
                        points:
                          type: integer
            text/csv:
              schema:
                type: string
        '401':
          description: Unauthorized
  /coupons/{token}:
    get:
      summary: Get coupon details
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Coupon found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coupon'
        '404':
          description: Coupon not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/{id}/wallet:
    get:
      summary: Get user wallet
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Wallet information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /users/{id}/redeem:
    post:
      summary: Request payout
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: integer
                  minimum: 1
                method:
                  type: string
                account:
                  type: string
              required:
                - amount
                - method
                - account
      responses:
        '202':
          description: Payout request accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  payout_id:
                    type: string
                  status:
                    type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...